rename: a new file for Curl_rename()
And make the cookie save function use it.
